requirements.txt
bing-image-downloader==1.1.2
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
duckduckgo_search==8.1.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
kaggle==1.7.4.5
kiwisolver==1.4.9
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
onnx==1.18.0
packaging==25.0
pillow==11.3.0
primp==0.15.0
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
